refactory the floder arch
