+++ lab3/.gitignore
@@ -3,3 +3,4 @@
 *.o
 .vscode
 venv
+build
