+++ lab3/include/mailbox.h
@@ -1,4 +1,4 @@
-#include "gpio.h"
+#include "peripherals/gpio.h"
 #include "tools.h"
 
 /* a properly aligned buffer */
