+++ lab3/src/tools.c
